# Removed decimal math from the numeric handler