… src/Npgsql/Properties/AssemblyInfo.cs
… src/Npgsql/TypeHandlers/NumericHandlers/NumericHandler.cs
… test/Npgsql.Benchmarks/NumericPerf.cs
… test/Npgsql.Tests/Types/NumericTypeTests.cs