@roji 2018-05-29
@YohDeadfall, @StillLearnin and anyone else, for extra credit we also have #448, which would allow reading/writing PostgreSQL `numeric` as .NET `BigInteger`, allowing handling values with an arbitrary number of digits.