@YohDeadfall 2018-05-29
> No time for a deep review, although the approach is very nice, with the unsafe casting of `decimal` to and from `DecimalRaw`.

Okay, I'm going to merge the PR. You can review it later, and I will push requested changes as a separate pull request. By the way, yesterday I realized how to make the new implementation a bit faster, but will do it for 4.0.1.

> Just to extra paranoid, I'm assuming that the internal memory structure of .NET `numeric` guaranteed by the runtime specs, so that we know for sure it will never change?

Yes, we can be sure that it will never change. It's a breaking change for .NET.